feat: rename defer to withResolvers
- update test and build system
- deprecated defer
- add withResolvers to match Promise.withResolvers

BREAKING CHANGE: defer is now deprecated in favor of withResolvers
