+++ lib/index.ts
@@ -1,2 +1,2 @@
-export { defer, Deferred } from './defer'
+export { defer, Deferred, withResolvers } from './withResolvers'
 export { deferGenerator, DeferredGenerator } from './deferGenerator'
