+++ lib/withResolvers.ts
@@ -1,4 +1,4 @@
-export function defer<T>(): Deferred<T> {
+export function withResolvers<T>(): Deferred<T> {
   let reject
   let resolve
   const promise = new Promise<T>((resolveFunc, rejectFunc) => {
@@ -12,8 +12,13 @@ export function defer<T>(): Deferred<T> {
   }
 }
 
+/**
+ * @deprecated With the release of [`Promise.withResolvers`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise/withResolvers) we've renamed `defer` to `withResolvers`.
+ */
 export interface Deferred<T> {
   promise: Promise<T>
   resolve: (value: T) => void
   reject: (error: Error) => void
 }
+
+export const defer = withResolvers
