+++ package.json
@@ -1,11 +1,17 @@
 {
   "name": "inside-out-async",
   "version": "0.0.0-development",
-  "description": "Promises and async iterators are awesome, lets turn them inside out",
+  "description": "Promises and async iterators are awesome, lets turn them inside out. Promise.withResolvers polyfill.",
   "types": "dist/index.d.ts",
   "exports": {
-    "require": "./dist/index.js",
-    "default": "./dist/index-esm.mjs"
+    "require": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    },
+    "import": {
+      "types": "./dist/index.d.mts",
+      "default": "./dist/index.mjs"
+    }
   },
   "main": "./dist/index.js",
   "engines": {
@@ -14,15 +20,16 @@
   "repository": "git@github.com:reconbot/inside-out-async.git",
   "homepage": "https://github.com/reconbot/inside-out-async",
   "scripts": {
-    "test": "mocha",
+    "test": "node -r esbuild-register --test lib/*-test.ts",
     "lint": "tsc && eslint lib/**/*.ts",
     "format": "eslint lib/**/*.ts --fix",
     "clean": "rm -rf dist-ts dist",
-    "build": "npm run clean && tsc -p tsconfig-build.json && rollup -c && ts-eager bundle-types",
+    "build": "npm run clean && tsc -p tsconfig-build.json && rollup -c && ts-eager bundle-types && cp dist/index.d.ts dist/index.d.mts",
     "prepublishOnly": "npm run build",
     "semantic-release": "semantic-release"
   },
   "keywords": [
+    "withResolvers",
     "inside-out-async",
     "defer",
     "differed",
@@ -32,30 +39,18 @@
     "async"
   ],
   "license": "MIT",
-  "dependencies": {},
   "devDependencies": {
-    "@microsoft/api-extractor": "^7.16.1",
-    "@types/chai": "^4.2.19",
-    "@types/mocha": "^8.2.2",
-    "@types/node": "^15.12.4",
-    "@typescript-eslint/eslint-plugin": "^4.27.0",
-    "@typescript-eslint/parser": "^4.27.0",
-    "chai": "^4.3.4",
-    "eslint": "^7.29.0",
-    "mocha": "^9.0.1",
-    "rollup": "^2.52.1",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "semantic-release": "^17.4.4",
-    "streaming-iterables": "^6.0.0",
-    "ts-eager": "^2.0.2",
-    "typescript": "^4.3.4"
-  },
-  "mocha": {
-    "bail": true,
-    "require": [
-      "ts-eager/register",
-      "chai/register-assert"
-    ],
-    "spec": "lib/**/*-test.ts"
+    "@microsoft/api-extractor": "7.42.3",
+    "@types/node": "20.11.25",
+    "@typescript-eslint/eslint-plugin": "7.1.1",
+    "@typescript-eslint/parser": "7.1.1",
+    "esbuild-register": "3.5.0",
+    "eslint": "8.57.0",
+    "rollup": "4.12.1",
+    "rollup-plugin-node-resolve": "5.2.0",
+    "semantic-release": "23.0.2",
+    "streaming-iterables": "8.0.1",
+    "tslib": "^2.6.2",
+    "typescript": "5.4.2"
   }
 }
