+++ tsconfig-build.json
@@ -1,6 +1,7 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
+    "target": "es2017",
     "outDir": "dist-ts",
     "noEmit": false,
     "module": "es6"
