+++ tsconfig.json
@@ -2,14 +2,14 @@
   "compilerOptions": {
     "noEmit": true,
     "allowJs": false,
-    "target": "es2017",
+    "target": "ES2022",
     "moduleResolution": "node",
     "module": "commonjs",
     "noImplicitAny": false,
     "strictNullChecks": true,
     "declaration": true,
     "importHelpers": true,
-    "forceConsistentCasingInFileNames": true
+    "forceConsistentCasingInFileNames": true,
   },
   "include": [
     "lib/**/*.ts",
